+++ src/angular-cesium-entities-editor/angular-cesium-entities-editor.module.ts
@@ -5,6 +5,8 @@ import { PolygonsEditorComponent } from './components/polygons-editor/polygons-e
 import { CirclesEditorComponent } from './components/circles-editor/circles-editor.component';
 import { CirclesEditorService } from './services/entity-editors/circles-editor/circles-editor.service';
 import { PolygonsEditorService } from './services/entity-editors/polygons-editor/polygons-editor.service';
+import { PolylinesEditorService } from './services/entity-editors/polyline-editor/polylines-editor.service';
+import { PolylinesEditorComponent } from './components/polylines-editor/polylines-editor.component';
 
 @NgModule({
   imports: [
@@ -14,14 +16,17 @@ import { PolygonsEditorService } from './services/entity-editors/polygons-editor
   declarations: [
     PolygonsEditorComponent,
     CirclesEditorComponent,
+		PolylinesEditorComponent,
   ],
   exports: [
     PolygonsEditorComponent,
     CirclesEditorComponent,
+    PolylinesEditorComponent,
   ],
   providers: [
     PolygonsEditorService,
     CirclesEditorService,
+    PolylinesEditorService,
   ]
 })
 export class AngularCesiumEntitiesEditorModule {