+++ src/angular-cesium-entities-editor/models/polygon-edit-options.ts
@@ -1,23 +1,9 @@
-import { CesiumEvent } from '../../angular-cesium/services/map-events-mananger/consts/cesium-event.enum';
-import { CesiumEventModifier } from '../../angular-cesium/services/map-events-mananger/consts/cesium-event-modifier.enum';
+import { PolylineEditOptions } from './polyline-edit-options';
 
-export interface PointProps {
-	color?: any;
-	outlineColor?: any;
-	outlineWidth?: number;
-}
 export interface PolygonProps {
 	material?: any;
 }
-export interface PolygonEditOptions {
-	addPointEvent?: CesiumEvent;
-	addPointModifier?: CesiumEventModifier;
-	addLastPointEvent?: CesiumEvent;
-	addLastPointModifier?: CesiumEventModifier;
-	removePointEvent?: CesiumEvent;
-	removePointModifier?: CesiumEventModifier;
-	dragPointEvent?: CesiumEvent;
-	dragShapeEvent?: CesiumEvent;
-	defaultPointOptions?: PointProps;
+
+export interface PolygonEditOptions extends PolylineEditOptions {
 	defaultPolygonOptions?: PolygonProps;
 }